Refactor CLI into separate module and files
