+++ src/hype/cli/commands/__init__.py
@@ -1,0 +1,4 @@
+from hype.cli.commands.run import run
+from hype.cli.commands.serve import serve
+
+__all__ = ["run", "serve"]
